+++ trunk/Muphic/Mu/source/src/Acorde.cpp
@@ -6,19 +6,19 @@ Acorde::Acorde() : Simbolo()
 	notas = new Nota[3];
 }
 
-Acorde::Acorde(float d) : Simbolo(d)
+Acorde::Acorde(int d) : Simbolo(d)
 {
 	notas = new Nota[3];
 }
 
-Acorde::Acorde(float d, int n1, int n2) : Simbolo(d)
+Acorde::Acorde(int d, int n1, int n2) : Simbolo(d)
 {
 	notas = new Nota[2];
 	notas[0].setTono(n1);
 	notas[1].setTono(n2);
 }
 
-Acorde::Acorde(float d, int n1, int n2, int n3) : Simbolo(d)
+Acorde::Acorde(int d, int n1, int n2, int n3) : Simbolo(d)
 {
 	notas = new Nota[3];
 	notas[0].setTono(n1);
@@ -27,7 +27,7 @@ Acorde::Acorde(float d, int n1, int n2, int n3) : Simbolo(d)
 }
 
 
-Acorde::Acorde(float d, int n1, int n2, int n3, int n4) : Simbolo(d)
+Acorde::Acorde(int d, int n1, int n2, int n3, int n4) : Simbolo(d)
 {
 	notas = new Nota[4];
 	notas[0].setTono(n1);
@@ -41,17 +41,18 @@ Acorde::Acorde(Nota* notas) : Simbolo()
 	this->notas = notas;
 }
 
+////////////////////////////////////////////////////////// WTF! lenght? //////////////////////////////////////////
 Acorde::Acorde(Nota notas[], int length) : Simbolo()
 {
 	this->notas = notas;
 }
 
-Acorde::Acorde(float d, Nota* notas) : Simbolo(d)
+Acorde::Acorde(int d, Nota* notas) : Simbolo(d)
 {
 	this->notas = notas;
 }
 
-Acorde::Acorde(float d, Nota notas[], int length) : Simbolo(d)
+Acorde::Acorde(int d, Nota notas[], int length) : Simbolo(d)
 {
 	this->notas = notas;
 }
